third_party/blink/web_tests/platform/webview/external/wpt/pointerevents/pointerevent_touch-action-pan-right-css_touch-expected.txt

This is a testharness.js-based test.
[FAIL] touch-action attribute test
  assert_greater_than: scroll x offset should be greater than 200 in the end of the test expected a number greater than 200 but got 200
Harness: the test ran to completion.
